+++ src/Elders.Cronus.Transport.RabbitMQ/Management/RabbitContractResolver.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json.Serialization;
+using System.Text.RegularExpressions;
+
+namespace Elders.Cronus.Pipeline.Transport.RabbitMQ.Management
+{
+    public class RabbitContractResolver : DefaultContractResolver
+    {
+        protected override string ResolvePropertyName(string propertyName)
+        {
+            return Regex.Replace(propertyName, "([a-z])([A-Z])", "$1_$2").ToLower();
+        }
+    }
+}