+++ src/Elders.Cronus.Transport.RabbitMQ/RabbitMqPipelineFactory.cs
@@ -1,5 +1,8 @@
-﻿using System;
+﻿using Elders.Cronus.Pipeline.Transport.RabbitMQ.Management;
+using Elders.Cronus.Pipeline.Transport.RabbitMQ.Management.Model;
+using System;
 using System.Collections.Concurrent;
+using System.Linq;
 
 namespace Elders.Cronus.Pipeline.Transport.RabbitMQ
 {
@@ -11,9 +14,12 @@ namespace Elders.Cronus.Pipeline.Transport.RabbitMQ
 
         private RabbitMqSession session;
 
-        public RabbitMqPipelineFactory(RabbitMqSession session, IPipelineNameConvention nameConvention)
+        private Config.IRabbitMqTransportSettings transportSettings;
+
+        public RabbitMqPipelineFactory(RabbitMqSession session, Config.IRabbitMqTransportSettings settings)
         {
-            this.nameConvention = nameConvention;
+            this.transportSettings = settings;
+            this.nameConvention = settings.PipelineNameConvention;
             this.session = session;
         }
 
@@ -21,6 +27,16 @@ namespace Elders.Cronus.Pipeline.Transport.RabbitMQ
         {
             if (!pipes.ContainsKey(pipelineName))
             {
+                var managmentClient = new RabbitMqManagementClient(transportSettings.Server, transportSettings.Username, transportSettings.Password, transportSettings.AdminPort);
+
+                if (!managmentClient.GetVHosts().Any(vh => vh.Name == transportSettings.VirtualHost))
+                {
+                    var vhost = managmentClient.CreateVirtualHost(transportSettings.VirtualHost);
+                    var rabbitMqUser = managmentClient.GetUsers().SingleOrDefault(x => x.Name == transportSettings.Username);
+                    var permissionInfo = new PermissionInfo(rabbitMqUser, vhost);
+                    managmentClient.CreatePermission(permissionInfo);
+                }
+
                 IRabbitMqPipeline pipeline = new UberPipeline(pipelineName, session, PipelineType.Headers);
                 pipeline.Open();
                 pipes.TryAdd(pipelineName, pipeline);