+++ RELEASE_NOTES.md
@@ -1,3 +1,13 @@
+#### 3.0.1 - 01.12.2016
+* Add setting for specifying admin port
+* Adds a Virtual Host, if such not present in RabbitMQ, and assigning permissions for the user passed to the initial settings 
+
+#### 3.0.1-beta0002 - 01.12.2016
+* Add setting for specifying admin port
+
+#### 3.0.1-beta0001 - 01.12.2016
+* Adds a Virtual Host, if such not present in RabbitMQ, and assigning permissions for the user passed to the initial settings 
+
 #### 3.0.0 - 08.09.2016
 * Adds support for IScheduleMessages sent to pipelines
 * Moves WithDefaultPublishersWithRabbitMq to Cronus